+++ src/lib/config.ts
@@ -60,7 +60,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "latest",
-    sdkVersion: "0.8.0",
-    genVersion: "2.409.0",
-    userAgent: "speakeasy-sdk/typescript 0.8.0 2.409.0 latest @lukehagar/discoursejs",
+    sdkVersion: "0.8.1",
+    genVersion: "2.409.8",
+    userAgent: "speakeasy-sdk/typescript 0.8.1 2.409.8 latest @lukehagar/discoursejs",
 } as const;